rename #include file nobody else should use
