+++ modules/c++/coda-oss.vcxproj.filters
@@ -957,7 +957,7 @@
     <ClInclude Include="hdf5.lite\include\import\hdf5\lite.h">
       <Filter>hdf5.lite</Filter>
     </ClInclude>
-    <ClInclude Include="xml.lite\include\xml\lite\xerces.h">
+    <ClInclude Include="xml.lite\include\xml\lite\xerces_.h">
       <Filter>xml.lite</Filter>
     </ClInclude>
     <ClInclude Include="hdf5.lite\include\hdf5\lite\H5_.h">
