+++ modules/c++/xml.lite/include/xml/lite/xerces_.h
